remove unused protocol
